chore: split migrations in multiple files
